USB: dwc3: debugfs: Add boundary check in dwc3_store_ep_num()
User can pass arguments as part of write to requests and endpoint number
will be calculated based on the arguments. There is a chance that driver
can access ep structue that is not allocated due to invalid arguments
passed by user. Hence fix the issue by having check and return error in
case of invalid arguments.

Bug: 28172322

Change-Id: I89e98b9fedbf7dc872d5d97ee081dea19aab511c
Signed-off-by: Vijayavardhan Vennapusa <vvreddy@qti.qualcomm.com>